+++ src/common/components/ContentCard/index.js
@@ -94,6 +94,7 @@ export type Props = {
   extraSubtitle?: any,
   forceUpdate?: boolean,
   children?: any,
+  inline?: boolean,
 };
 
 const ContentCard = ({